Renamed model_repository directory to model_register Changed model predict method interface to not delegate inference_id management for the model
	Model predict method now returns a tuple containing a list of floats (the model output)
	and a string (model diagnosis)
